Finalized main game loop, completed core scoring system and end screen.
Finished the main gameplay loop, ensured that the text displayed was not the same as the color. Implemented an end/restart screen where the player can replay the game, go to the main menu and quit the application. Also implemented a time-based scoring system which rewards the player with more points the faster they guess the correct answer.