Fixed color repetition, timer decimal places, added description of game
Fixed an issue where colors would repeat throughout the rounds, the colors no longer do this and makes the game more difficult. Changed how the timer is displayed to the player, now only displays the timer to 2 decimal places. Added a brief summary/description of the game in the Main Menu and adjusted some visual aspects. Started testing particles for visual feedback for player.